requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# AI & NLP
openai==1.3.7
whisper==1.1.10
spacy==3.7.2
transformers==4.35.2

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Document Processing
reportlab==4.0.7
python-docx==1.1.0
pypdf==3.17.1
jinja2==3.1.2

# HTTP & File Handling
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
websockets==12.0
starlette==0.27.0

# Audio Processing
librosa==0.10.1
soundfile==1.0.1
pyaudio==0.2.11
webrtc-vad==2.0.10
numba==0.58.1

# Security & Compliance
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
